Implement optifine exclusion
